Move tests outside src
